requirements.txt
# Core dependencies
streamlit
requests
python-dotenv

# AI integration
openai

# Utilities
python-whois

# Development tools